Improve dataset declaration
Output datasets in VR are declared by giving them a name, a unit
information object, and a storage data type (most of them use the
Double_t type). These three values are then stored in different arrays
where they are collected in order.

The way VR declared these datasets was a bit confusing though, mostly
because there was a lot of code duplication, and because the addition of
values into the different arrays (names, units, storage data types)
advanced at different rates. This made it difficult, for example, to
understand the data type or unit information associated to a dataset
name.

This commit improves this situation dramatically. Instead of advancing
at different rates, now using a single declaration a dataset is fully
specified in terms of name, unit information and storage data type. Upon
this simpler specification we also build a number of utility methods
that perform several declarations at once (e.g., for "<name>{x,y,z}"
dataset trios). The end result is a code base with less duplication,
and easier to follow and maintain.

As part of this change I actually realised some mass-related datasets
did not have the proper unit associated to them, hence I fixed them.
There was also a typo in one set of aperture dataset declarations
("Aperture_veldips_gas_sf_" instead of "disp") so I fixed that as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>