Simplify highT loop structure
Many of the conditions checked within this loop are constant throughout
all the elements of the loop, so it makes sense to move them outside of
it. The loop itself also doesn't need to be within an 'if' condition
that is checked inherently the first time the loop runs.

Finally, since I'm touching so much code, I've fixed the indentation of
this whole block.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>